fix(types): add PolygonPressEvent type
