+++ src/sharedTypes.ts
@@ -99,20 +99,3 @@ export type MarkerPressEvent = NativeSyntheticEvent<{
    */
   position?: Point;
 }>;
-
-export type PolygonPressEvent = NativeSyntheticEvent<{
-  action: 'polygon-press';
-  /**
-   * @platform iOS: Google Maps
-   */
-  id?: string;
-  /**
-   * @platform iOS: Apple Maps
-   * @platform Android
-   */
-  coordinate?: LatLng;
-  /**
-   * @platform Android
-   */
-  position?: Point;
-}>;
